CSN: CSN-platforms-29
The beginning of "productizing" the PCIE controller.

In this checkin the existing structure and modules are not modified, but new
things are added so that we can build them up to the organization that we want
while existing models continue to function.

Includes:
 * Extremely preliminary version of the PCIE VHDL.
 * Initial import of PCIE VHDL into BSV
 * Initial set of Physical Devices:
    * LEDs
    * Switches
    * Buttons
 * Initial Physical Platforms
    * HTG PCI Enabled
    * HTG PCI Disabled

Open question: Who instantiates the physical platform and maps the virtual
devices onto it?

Michael P.