CSN: CSN-platforms-129
Used actual Xilinx Bram ports, which is basically 2-ports, each port can either be used as a read or as a write at a time. The interface remains the same. So now, if one instantiates a n-readPort, 1-writePort bram, you use only ceil(n/2) brams. But the downside is, write to the second port will conflict with the read to the second port, so they can not be written on the same rule.