CSN: CSN-platforms-155
The round robin arbiter had a lot of logic depth that was slow to compile in
Bluespec and generated difficult place & route problems.  After some
experimenting, found an implementation that compiles quickly and, I think,
is equivalent.

This code has a chain through all arbitrated bits.  We'll have to switch to
a tree if the arbitration set grows too large.