CSN: CSN-platforms-200
Updated clocking methodology for all FPGA designs. "Clocks" is now a new
physical device exported by each Physical Platform. Three implementations
of clocks device have been provided:

  - a single-ended device that derives a Clock object from a single top-level
    wire CLK,
  - a differential device that derives a Clock object from a pair of wires
    CLK_P and CLK_N, and
  - a simulation device that instantiates a simulated oscillator.

For real hardware platforms, the wires are *not* bound to UCF pins within
the device implementation; the binding happens at the Physical Platform
layer, so that multiple Physical Platforms can share the same single-ended
and differential clock modules.

Clock-multiplication for the model clock is now part of the Clocks device.
The crystal clock frequency is a local AWB parameter, but the multiplication
and division factors are exported as global parameters so that different
models can use different factors.

Each Clocks implementation also exports a Reset object. The Clock and Reset
objects together form the "driver" part of the physical device.

The Clocks device also internally implements the logic for generating a soft
reset, and ties the effect of the soft reset into the main Reset signal it
exports as part of the driver. However, the soft reset needs to be explicitly
*triggered* elsewhere. The clocks device therefore also exports a
SOFT_RESET_IFC interface up to the Platform, which can then pass it into
whichever Physical Device is capable of actually triggering the Soft Reset on
the particular platform.