CSN: CSN-platforms-303
Add a functional, internally stateless, implementation of a FIFO.  Functions
take the state of the FIFO as a parameter.  Those that update the FIFO return
the new state.  The state is simply a struct that can be stored in any memory.

Clearly the FIFO has different semantics than usual, since reading and writing
from different rules in the same cycle is impossible.  The code simply manages
a ring buffer.  The most likely use is in algorithms that need a large array
of relatively small FIFOs.