CSN: CSN-platforms-323
Add build of .ba files to the set of objects built as a side effect of building
the .bo files.  I suspect this should fix the problems Elliott was trying
to fix up at the top level SCons script.  Elliott's most recent changes
are unstable, with regressions in Hudson failing.  This change fixes that
instability, since SCons now knows it built the .ba files.