CSN: CSN-platforms-325
Fixed the conversion module to prevent more than one outstanding read at a time.  This was causing a number of bugs in the PHY, including overwhelming its limit of 4 outstanding transactions, and I think causing some issue in back to back addrAck's. This should be fixed at some point in an effort to mine out maximum parallelism.