CSN: CSN-platforms-390
Changed name of model to 'XUP Virtual Devices' to eliminate duplication 
with name 'Standard Virtual Devices' used in std-virtual-device.apm.

Only difference in this submodel from 'Standard Virtual Devices' seems
to be use of a NULL central cache instead of a local memory central
cache. There also seemed to be no usage of this submodel.