CSN: CSN-platforms-423
Tweak the streaming read code for uncached scratchpads to behave the way it
should have in the first place.  The old code didn't record a read line until
the data was returned by the host.  By then it was too late, since all
streaming requests for a line are most likely generated before the line
is returned from host memory.

The new code tracks the most recent read line request.  If the new request
matches the old line, the request bypasses the host memory pipeline and
is returned from a shared buffer once the data arrives.