CSN: CSN-platforms-449
Functional JTAG Physical Channel for Xilinx devices.  This can be used
on any Virtex chip for activities such as debugging faster physical
channels.  The speed is low (~100B/s), but can probably be improved
with a little effort. 

XMD is a pretty terrible tool.  For the JTAG to operate, it requires
frequent user input in the form of carriage returns.  It also requires
the hardware to write at least 32 bytes before XMD will deign to read
its buffer.  Just another example of a fine Xilinx tool.