CSN: CSN-platforms-461
Added a few build tools, most notably the mcd build tool, which
injects ucf constraints if MCD codes are detected.  Eventually, we
should make this handle all known synchronization primitives, but it
currently does not.  

Also added new XST build which does not invoke the basicio script on
the top level verilog.  This points to a reorganization, making the
basicio invocation a first class citizen.