CSN: CSN-platforms-471
This large checkin (I seem to be fond of these lately) refactors large
portions of the build pipeline code.  This change has been anticipated
for some time, but the needs of the multifpga build branch have
finally required it to happen.  This code is also needed for my
attempt to support the fish-class simulators.  The basic idea is that
all the functionality that was in SCons.top has been logically
partitioned into sub builds.  The only thing which has not been
changed are the sw and hw sub-scons.  These will eventually turn into
SCons functions, finally enabling us to get the build deps right (as
opposed to the current amorphous blob-o-deps we get from the sub-scons).