CSN: CSN-platforms-481
Intermediate check-in of PCIE support for XUPV5.  This code does
appear to scale to other Xilinx boards, something that might be useful
in the future.  Also included is a minor rework of the build pipeline
which raises our compatibility with pygraph to version 1.8.  This was
needed because earlier versions of pygraph can no longer be built, but
it does break backwards compatibility.  This should be okay moving
forward, but I would like to have an automatic way of building a
python enviroment.