CSN: CSN-platforms-489
Resurrected the altera jtag by reverting to old driver code.  Not a
great solution, but I don't have time to investigate why the combined
driver is not working.  Added support for DE2, DE3 boards.  I haven't
tested yet as the DE2 is not hooked up and I haven't got the Synplify
parameters for the DE3.