CSN: CSN-platforms-497
Added xupv5 serial device.  Since serial is a general device, I
reorganized it to depend on a platform-specific ucf.  Some more work
is needed.  1) we need to ensure a constant 100Mhz clock - this is
not an issue for XUPv5 and XUPv2, but other platforms may not have an
available 100Mhz oscillator.  2) I need more robust flow control
enable/disable as some platforms, including XUPv5, do not support HW
flow control.