CSN: CSN-platforms-524
Moved locatestate to /var when packaging for Ubuntu. Removed 
requirement for chkconfig which does not seem to work properly on 
Ubuntu since they moved to using upstart. So there are two
things to note:

1) When one wants to manually install in /usr one nieds to use 
the following configure command:

 % ./configure --prefix=/usr --sysconfdir=/etc/ --localstatedir=/var \
                --mandir=/usr/share/man --infodir=/usr/share/info

I need to figure out if there is a standard way for autoconf to set
these dirs when prefix=/usr.


2) You need to manually configure the init script /etc/init.d/init-fpga-ctrl.
On SUSE one can use chkconfig... On Ubuntu it is TBD.