CSN: CSN-leap-platforms-545
Remove newline from end of file. This was breaking things when the XST build
pipeline appended flags, since the extra newline caused them to be interpreted
as a new set of commands, rather than an append to the existing set of commands.

This is more of a workaround. The real fix would be to have the XST pipeline
stage chomp spurious newlines out from the end of the file before appending.