CSN: CSN-leap-platforms-556
Updated XUPV5 run script to do a more intelligent interrogation of the
PCIE device.  It can now figure out the magic PCIE memory settings
without user intervention, and can prompt the user to reboot computer
if the PCIE is not properly configured.  

It cannot load or unload the PCIE device driver, though that
functionality will be coming soon.