CSN: CSN-leap-platforms-562
Backport

CSN: CSN-leap-platforms-555
          
Improved the behavior of leap-fpga-ctrl with respect to the optional control scripts that
live in <PREFIX>/share/leap/scripts, which do extra operations on hardware devices for
certain leap-fpga-ctrl commands. Now if you want to use such a script you must enable
it explictly with a line in the configuration file in <PREFIX>/etc/leap/config. In
specific, one must have a line of the following form for each device in that file:

has_control_script = yes|no

In order to avoid silent failures all devices must now have such a line in their 
section of the configuration file. In addition, if the device is specified to have 
a script and it does not exist, then leap-fpga-ctrl will abort with an error message.

Fixed debian installation so leap-fpga-ctrl is actually installed suid root.