CSN: CSN-leap-platforms-574
Added several high-speed serial implementations for the XUPV5.  These
GTP/GTX links are generally available on Xilinx chips and boards, and
are very useful for board to board communication in the context of a
multiple FPGA implementation.  Although what I have implemented is
device specific, in practice porting to different links is
straightforward (indeed this commit includes two different links on
the XUPV5).  Also good is that the links seem to be scaling with
Moore's law in bandwidth and latency.