CSN: CSN-leap-platforms-649
Force synthesis of the PCIe driver and BlueNoC bridge as separate boundaries
using (* synthesize *).  This guarantees that the entire module is part of
the designated area group.  Without the boundary some compiler-generated
temporary names wound up not matching the INST pattern and were forced outside
the area group, leading to timing problems.

The unfortunate hack here is that pcie-virtex6-include.v is needed to include
the generated Verilog.  Ideally, we would have syntax in AWB or automatic
detection of generated Verilog files as part of the build process.