CSN: CSN-leap-platforms-670
1) Refactored the aurora code to cleanly separate out the verilog
driver from the upper portions of the stack.  It may be that as we add
support for the ML707 that we will want to have a board-specific ucf
module as well.  We could also do the same for the 5 and 6.125 Gbps
transceivers.  I did not attempt to port the Virtex 5 devices although
that may be a good idea.

2) Fixed a few bugs in the aurora verilog.  Xilinx thoughfully reset
the aurora block every few thousand cycles.  I also changed some gtx
signal parameters to match the IBERT test bench.