CSN: CSN-leap-platforms-671
1) Added icon and ila ngcs.  These are the components of the chipscope
logic analyzer.  Currently, they are only used in xilinx-provided
verilog as a low-level debugging aide. However, this sort of global
resource is the kind of thing that is screaming for inclusion as a
soft-service.

2) Minor modifications and a flow control bug fix to the aurora flow
control layer.  The aurora link is quite stable now, and I have
sucessfully run a partitioned HAsim model on the ML605 setup.