CSN: CSN-leap-platforms-675
Changed the ML605 to have four additional Aurora SERDES channels.
These channels are attached to the pinned out hpc connector.  

Unfortunately, due to clock management issues, we cannot yet make use
of all of the available SERDES.  Each SERDES appears to require its
own MMCM if it is to be used individually.  It may be possible to
relax this constraint, but I am not sure how to do it yet.

Fortunately, the SERDES can be ganged to for a wider, multi-lane
channel.  In this configuration, they can share a single MMCM.