CSN: CSN-leap-platforms-677
Conversion from double to single took the resulting 32 bit single precision
value and remapped it to be stored as a double, which is what Alpha wants.
The Alpha pipeline took the remapped value and did the same thing, resulting
in garbage.  Removed the remapping here, since this code is supposed to
be ISA neutral.  Double to single now returns data in the low 32 bits.