CSN: CSN-leap-platforms-679
1) Added support for several different transceiver throughputs.  

2) Refactored the SERDES flowcontrol layer to support different clock
ratios by parameter.

3) Simplified some program logic to achieve 200MHz timing closure.  I
don't think it's likely that we'll be able to get beyond 200MHz, at
least on the ML605.  

Things that need to be done:

1) Better code factoring.  Currently the aurora device ucf is
hard-coded for 4gbps.  It would be nice if that were not the case.

2) We're only using the 16-bit interface, and 5gbps, which requires
that the flow control logic operates at 250MHz, is really not
acheiveable.  However, Xilinx has a 32-bit interface which might allow
us to get to 8gbps/lane.  

3) Better area groups.  I took a stab at areagroups, and things seem
to work, but it needs to be tightened up.