CSN: CSN-leap-platforms-680
Add a new LUTRAM that is initialized by a Get interface passed in as a
function.  With this it is easy to initialize a LUTRAM from software by
tying an RRR stream (morphed into a Get) directly to a RAM.  When we
finally drop RRR and replace it with soft connections toGet() will be
defined automatically, making initialization from software a one-line
declaration.