CSN: CSN-leap-platforms-682
1) Added support for different serdes oscillator frequencies.  This
allows us to use the XM104 board in conjunction with other boards and
the SMA based SERDES on the ML605.  

2) Added some new data structures to describe these changes.

We can now run multifpga implementations with >2 FPGAs!

TODO:

1) We need to handle different clock frequencies and widths in the
same manner that we handle the oscillator frequencies. This will
reduce the number of awb and files in the SERDES library.