CSN: CSN-leap-platforms-686
We are seeing corruption during programming if two devices are programmed at the
same time.  Change leap-fpga-ctrl to permit only one device to be in the
programming state at a time.  No script changes are required, since other
requests block until the lock is released.

The --force option ignores the lock.