CSN: CSN-leap-platforms-693
1) Fixed potential (though very contrived) buffer overflow
vulnerability in leap-fpga-ctrl.  Now all buffer are sized according
to the maximum size allowed by iniparser, and this length is checked
when accessing the configuration database. 

2) Improved inter-fpga handshaking.  Aurora endpoints can now only
initialize if the partner Aurora is, itself, not initialized.  This
helps us deal with FPGAs which do not have the clean reset control of
the LEAP PCIE interface.

This implementation is a degenerate implementation of a more general
epoch-based reset distribution mechanism.