CSN: CSN-leap-platforms-700
Initial revision of a frame-based reliability layer for inter-FPGA
links.  This serves as a drop-in replacement for the existing
flowcontrol layer.

Current verion implements a sliding window-based protocol with acks.
Some features are missing.  Retransmission is currently disabled, and
the checksum function is constant.  We also lack a self-healing
function in the case of multiple errors.  Furthermore, we ignore error
information propagated by the SERDES phy.  All of these issues will be
addressed in a subsequent revision.  From a performance and area
perspective, however, the code is largely complete.