CSN: CSN-leap-platforms-705
1. Fixed deadlock issue in reliable inter-fpga procotol.  If a partial
packet containing a data dependency is corrupted, retransmission will
not occur because the packet never finishes.  To fix this we begin the
timeout at the start of the packet, so that corrupt partial packets
will eventually be retransmitted. 

2.  Added bluenoc-ioctl, which allows scripts to interface with the
bluenoc device driver.  This code, coupled with changed to the
platform-specific scripts, subsumes the old soft-reset interface. It
also gives the platform script precise control over the reset behavior
of the FPGA, which is necessary for multifpgas.