CSN: CSN-leap-platforms-706
1)Added parameter to aurora devices which controls the length of the
intialization phase.  Long initialization phases seem to improve the
reliability of some of the aurora interfaces, but they are burdensome
in simulation.

2) Added a reliable aurora protocol with intial handshaking.  In pure
LEAP PCIE platforms, handshaking is not needed because we have
software control of reset through the PCIE interface.  However, some
platforms may not have this level of control and therefore require a
handshaking scheme when coming out of reset.