CSN: CSN-leap-platforms-707
Yet another refinement to the reliable inter-fpga transport protocol.
This version leverages the clock differential between the user and
aurora domains to reduce the cost of the crc.  CRC is calculated over
the user-domain word, rather than the aurora domain word, which is
several times smaller.  Throughput is around 500MB/s, though I suspect
it can be improved.