CSN: CSN-leap-platforms-711
Added a write-bypass unguarded BRAM.  If Xilinx configures the BRAM to
be write-first, then simulataneous reads and writes to the same
address result in the wrong read value.  This bram adds a
register-based bypass to minimize the amount of scheduling in the
unguarded BRAM.

We should eventually investigate the write-first mode fully.